+++ lib/test/apyfixedarray/test_arithmetic.py
@@ -517,11 +517,6 @@ def test_array_div(fixed_array: type[APyCFixedArray]):
         )
     )
 
-    # Does not die
-    a = fixed_array.from_float([5], bits=100, int_bits=50)
-    b = fixed_array.from_float([0], bits=250, int_bits=100)
-    assert (a / b).is_identical(fixed_array([0], bits=351, int_bits=201))
-
 
 @pytest.mark.parametrize(
     ("fixed_array", "fixed_scalar"),
@@ -1443,21 +1438,21 @@ def test_two_limb_division_result_32_one_long():
 
 @pytest.mark.parametrize("int_bits_1", [10, 30, 50, 100, 300])
 @pytest.mark.parametrize("int_bits_2", [10, 30, 50, 100, 300])
-def test_arrays_dont_crash_on_zero_div(int_bits_1: int, int_bits_2: int):
-    """
-    In APyTypes, array division-by-zero results in an undefined values. However,
-    APyTypes should not crash when such a division occurs
-    """
+def test_arrays_raises_on_zero_div(int_bits_1: int, int_bits_2: int):
     a = APyFixedArray.from_float([0.0, 1.0, 2.0], int_bits=int_bits_1, frac_bits=0)
     b = APyFixedArray.from_float([0.0, 0.0, 1.0], int_bits=int_bits_2, frac_bits=0)
-    _ = a / b
-    _ = b / a
+    with pytest.raises(ZeroDivisionError):
+        _ = a / b
+    with pytest.raises(ZeroDivisionError):
+        _ = b / a
 
 
 @pytest.mark.parametrize("int_bits_1", [10, 30, 50, 100, 300])
 @pytest.mark.parametrize("int_bits_2", [10, 30, 50, 100, 300])
-def test_arrays_dont_crash_on_zero_div_with_scalar(int_bits_1: int, int_bits_2: int):
+def test_arrays_raises_on_zero_div_with_scalar(int_bits_1: int, int_bits_2: int):
     a = APyFixedArray.from_float([0.0, 1.0, 2.0], int_bits=int_bits_1, frac_bits=0)
     b = APyFixed.from_float(0.0, int_bits=int_bits_2, frac_bits=0)
-    _ = a / b
-    _ = b / a
+    with pytest.raises(ZeroDivisionError):
+        _ = a / b
+    with pytest.raises(ZeroDivisionError):
+        _ = b / a